+++ JFramework/nbproject/private/private.properties
@@ -1,0 +1,4 @@
+deploy.ant.properties.file=/home/johan/snap/netbeans/76/tomcat90.properties
+j2ee.server.home=/usr/share/tomcat9
+j2ee.server.instance=tomcat90:home=/usr/share/tomcat9
+user.properties.file=/home/johan/snap/netbeans/76/build.properties
