+++ JFramework/nbproject/private/private.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
+    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
+        <group>
+            <file>file:/home/johan/Documents/Programmation/Naina/Framework/JFramework/src/java/etu1933/framework/view/ModelView.java</file>
+            <file>file:/home/johan/Documents/Programmation/Naina/Framework/JFramework/src/java/etu1933/framework/servlet/FrontServlet.java</file>
+            <file>file:/home/johan/Documents/Programmation/Naina/Framework/JFramework/web/WEB-INF/web.xml</file>
+            <file>file:/home/johan/Documents/Programmation/Naina/Framework/JFramework/src/java/etu1933/framework/Mapping.java</file>
+            <file>file:/home/johan/Documents/Programmation/Naina/Framework/JFramework/web/index.html</file>
+        </group>
+    </open-files>
+</project-private>
