+++ JFramework/src/java/etu1933/framework/Mapping.java
