+++ JFramework/src/java/etu1933/framework/view/ModelView.java
@@ -1,0 +1,16 @@
+package etu1933.framework.view;
+
+public class ModelView {
+    String view;
+
+    public String getView() {
+        return view;
+    }
+
+    public ModelView setView(String view) {
+        this.view = view;
+        return this;
+    }
+
+    public ModelView() {}
+}
