+++ JFramework/web/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<!--
+Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Html.html to edit this template
+-->
+<html>
+    <head>
+        <title>TODO supply a title</title>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    </head>
+    <body>
+        <div>
+            <form method="post" action="/">
+                <input type="text" value="Url sending">
+                <input type="submit" value="post">
+                <a href="/?envoie=coucou&envoie2=monsieur"><input type="button" value="get"></a>
+            </form>
+        </div>
+    </body>
+</html>
