--- README.md
@@ -1,2 +1,0 @@
-# JFramework
-Placer le source des classes ou on veut avoir la liste des fonctions par ces classes
