+++ Temporaire/nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=5e01697a
+build.xml.script.CRC32=304ed1f1
+build.xml.stylesheet.CRC32=1707db4f@1.93.0.1
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=5e01697a
+nbproject/build-impl.xml.script.CRC32=c8ea0aca
+nbproject/build-impl.xml.stylesheet.CRC32=334708a0@1.93.0.1
